Public interface

Index

GENeSYSMOD.CSVResult

GENeSYSMOD.NoInfeasibilityTechs

GENeSYSMOD.NoRawResult

GENeSYSMOD.OneNodeSimple

GENeSYSMOD.OneNodeStorage

GENeSYSMOD.TXTResult

GENeSYSMOD.TXTandCSV

GENeSYSMOD.TwoNodes

GENeSYSMOD.WithInfeasibilityTechs

GENeSYSMOD.fetch_data_release

GENeSYSMOD.genesysmod

GENeSYSMOD.genesysmod_build_model

GENeSYSMOD.genesysmod_build_model_dispatch

GENeSYSMOD.genesysmod_dispatch

GENeSYSMOD.update_and_process_data

Types

GENeSYSMOD.CSVResult — Type

CSVResult - Raw results will be printed in CSV format. Each variable will have its CSV file named after the variable.

GENeSYSMOD.NoInfeasibilityTechs — Type

NoInfeasibilityTechs - type to use when the infeasibility techs are not needed.

GENeSYSMOD.NoRawResult — Type

NoRawResult - The raw results will not be printed.

GENeSYSMOD.OneNodeSimple — Type

OneNodeSimple - The dispatch is conducted for the considered_region in isolation from bordering regions.

GENeSYSMOD.OneNodeStorage — Type

OneNodeStorage - The dispatch is conducted for the considered_region while the interaction with the rest of the system is modeled as storages.

GENeSYSMOD.TXTResult — Type

TXTResult - Raw results will be printed in TXT format. A single txt file containing all values will be printed. This type as an argument filename used to define the name to use for the txt file. By default the argument takes the value "results".

GENeSYSMOD.TXTandCSV — Type

TXTandCSV - Raw results will be printed in both TXT and CSV formats.

GENeSYSMOD.TwoNodes — Type

TwoNodes - The dispatch is conducted for the considered_region and another dummy region consisting of an aggregation of the other regions.

GENeSYSMOD.WithInfeasibilityTechs — Type

WithInfeasibilityTechs - type to use when infeasibility techs are desired in the model. Typically they are used to help troubleshoot infeasibility as they provide a relaxation that allow the model to solve and to find the origin of theinfeasibility in the results.

Methods

GENeSYSMOD.fetch_data_release — Method

fetch_data_release(; version_tag::String="latest", dataset_name::String="MiddleEarth", dest_dir::AbstractString=pwd())

Download GENeSYS-MOD input data files from a GitHub release of the data repository.

Arguments

version_tag::String: The version tag of the data release to download. Defaults to "latest", which uses LATEST_DATA_VERSION.

dataset_name::String: The name of the dataset to download. Defaults to "MiddleEarth".

dest_dir::AbstractString: The destination directory where the files will be saved. Defaults to the current working directory.

Downloads

RegularParameters_<dataset_name>.xlsx: Regular parameters file for the specified dataset

Timeseries_<dataset_name>.xlsx: Timeseries data file for the specified dataset

Errors

Throws an error if either file fails to download, including the URL and error details.

Examples

fetch_data_release(version_tag="v1.0.5", dataset_name="Europe_EnVis_Green", dest_dir="path/to/save")

GENeSYSMOD.genesysmod — Method

Run the whole model. It runs the whole process from input data reading to result processing. For information about the switches, refer to the datastructure documentation.

GENeSYSMOD.genesysmod_build_model — Method

Build the model without running it. For information about the switches, refer to the datastructure documentation.

GENeSYSMOD.genesysmod_build_model_dispatch — Method

Build the dispatch model. A previous run is necessary to allow to read in investment decisions. For information about the switches, refer to the datastructure documentation

GENeSYSMOD.genesysmod_dispatch — Method

Run the simple dispatch model. A previous run is necessary to allow to read in investment decisions. For information about the switches, refer to the datastructure documentation

GENeSYSMOD.update_and_process_data — Method

update_and_process_data(; settings_file=nothing, scenario_option="Europe_EnVis_Green",
                        datadir=nothing, resultdir=nothing, output_format="long", processing_option="both")

Fetches the latest changes from the GENeSYS_MOD.data repository, merges them, and processes the data using a Python function.

Arguments

settings_file::Union{String, Nothing}: Absolute path to the settings file. If nothing, defaults to "Setfilterfile.xlsx" in the current directory.

scenario_option::String: Scenario option for the processing function. Default is "EuropeEnVisGreen".

datadir::Union{String, Nothing}: Directory of the data repository. If nothing, defaults to "GENeSYS_MOD.data" in the current directory.

resultdir::Union{String, Nothing}: Directory to save the output files. If nothing, the files are not moved and can be found in GENeSYS_MOD.data/Output.

output_format::String: Format of the output data, "long" or "wide". Default is "long".

processing_option::String: Specifies what data to process. Options are "parametersonly", "timeseriesonly", or "both". Default is "both".

debugging_output::Bool: Specifies if there should be more information in case of a code error. Default is false.

data_base_region::String: Specifies the base region using a country code. If values are missing in other regions, the base region's values will be used as fallback. Default is "DE".

Returns

result::Bool: true if the process completes successfully.

Description

Fetches the latest changes from the GENeSYS_MOD.data repository.

Attempts to merge the changes up to 3 times.

Calls the Python master_function to process the data.

Moves the output files to the specified resultdir if provided and based on the processing_option.

Returns true if the process completes successfully.

Example

result = update_and_process_data(settings_file="path/to/Set_filter_file.xlsx", scenario_option="Test_Scenario",
                                 datadir="path/to/data", resultdir="path/to/save", output_format="wide",
                                 processing_option="parameters_only")